+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Enhancements
 - Keyword arguments to `DBNStore.to_parquet` will now allow `where` and `schema` to be specified
+- Improve records processing time
 
 #### Bug fixes
 - Fixed an issue where validating the checksum of a batch file loaded the entire file into memory