+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-## 0.43.1 - TBD
+## 0.43.1 - 2024-10-15
 
 #### Enhancements
 - Keyword arguments to `DBNStore.to_parquet` will now allow `where` and `schema` to be specified
-- Improve records processing time
+- Improved record processing time for the `Live` client
 
 #### Bug fixes
 - Fixed an issue where validating the checksum of a batch file loaded the entire file into memory